feat(chrome): Windy-style scrubber pill
The docked bottom bar becomes a floating pill: transport, LIVE badge and clock
on one row, a drawn time track on the next. The track carries one tick per
volume, hour labels in the radar's own timezone, the forecast tail and the live
loop window, and scrubs on click or drag.

hide_toolbar dies with the dock, along with its palette action and registry row.
